% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{api_key}
\alias{api_key}
\title{Find or define API key for cortical.io}
\usage{
api_key(key = NULL)
}
\arguments{
\item{key}{"quoted" key or by default searches in .Renviron en Sys.setenv}
}
\value{
key
}
\description{
This function is not for direct use.
}
