% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{cortical_expressions}
\alias{cortical_expressions}
\title{Expression}
\usage{
cortical_expressions(text, sparsity = 1, retina_name = "en_associative",
  api_key = NULL)
}
\arguments{
\item{text}{input a character vector:  "a small lamb walked ..."}

\item{sparsity}{standard 1.0, measure of sparsity.}

\item{retina_name}{one of en_associative or en_synomous see retina.}

\item{api_key}{your key, if you saved the key in .renviron leave this NULL.}
}
\description{
This method returns a retina representation (a Fingerprint) of the
result of the input expression. An expression can contain Terms,
Fingerprints, Texts and even nested expressions. The expression must be
passed in the body of the method call. The sparsity parameter can be used
to sparsify the evaluated expression to a given percentage. The parameter
is only interpreted when in the range (0,1), and only applied, if the
fingerprint is more dense than the desired sparsity level.
\url{http://api.cortical.io/Expression.htm#!/expressions/resolveExpression_post_0}
}
