% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{cortical_image}
\alias{cortical_image}
\title{Return visualisation of expression}
\usage{
cortical_image(text, retina_name = "en_associative", image_scalar = 2,
  plot_shape = "circle", image_encoding = "base64/png", sparsity = 1,
  api_key = NULL)
}
\arguments{
\item{text}{the search}

\item{retina_name}{one of en_associative or en_synomous see retina.}

\item{image_scalar}{defaults to 2}

\item{plot_shape}{circle or square}

\item{image_encoding}{defaults to base64/png}

\item{sparsity}{defaults to 1}

\item{api_key}{your key, if you saved the key in .renviron leave this NULL.}
}
\description{
Implementation Notes
This method returns a visualization of the specified expression.
The expression must be passed in the body of the method call. The
image_scalar parameter defines the scale of the output image.
}
\details{
The plot_shape parameter defines the shape of the output image.
If this parameter is unspecified, (null), then the default circle will be
used. The image_encoding parameter specifies the type of image encoding used
(in this UI it can only be displayed using "base64/png").

The sparsity parameter can be used to sparsify the evaluated expression to a
given percentage. The parameter is only interpreted when in the range (0,1),
and only applied, if the fingerprint is more dense than the desired sparsity
level. The following simple example can be used to get an image for a single
input term.
}
