% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina.R
\name{cortical_retina}
\alias{cortical_retina}
\title{Return information about the Retina}
\usage{
cortical_retina(retina_name = NULL, api_key = NULL)
}
\arguments{
\item{retina_name}{one of NULL,  en_associative or en_synomous see retina.}

\item{api_key}{your key, if you saved the key in .renviron leave this NULL.}
}
\value{
information about the retina (database of words)
}
\description{
This function is not that useful, it returns basic information about the
current state of the two retinas in use.
}
