% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{cortical_terms}
\alias{cortical_terms}
\title{Find word in retina}
\usage{
cortical_terms(term, retina_name = "en_associative", start_index = 0,
  max_results = 10, get_fingerprint = FALSE, api_key = NULL)
}
\arguments{
\item{term}{the search word that you would like to search for}

\item{retina_name}{one of en_associative or en_synomous see retina.}

\item{start_index}{start of the index (if multiple pages are returned)}

\item{max_results}{maximum number of result per page.}

\item{get_fingerprint}{true or false, return sparse matrix representation}

\item{api_key}{your key, if you saved the key in .renviron leave this NULL.}
}
\value{
a description of the word in the retina.
}
\description{
This function returns what the term means for the retina.
Results include df, score, and position types such as adjective, or noun.
}
\section{Advanced Search}{

This endpoint returns a term object with meta-data for an exact match, or
a list of potential retina terms if the string contains one or more of the
wildcard characters, '*' and '?'. The wildcard characters must be initially
preceded by at least 3 characters. The asterisk wildcard, '*', represents
zero or more characters. The question mark wildcard, '?', represents
exactly one character.
Information copied straight from the website \link{http://api.cortical.io/}
}
\seealso{
Other terms: \code{\link{cortical_similar_terms}},
  \code{\link{cortical_terms_context}}
}

