% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{cortical_terms_context}
\alias{cortical_terms_context}
\title{Find context where this word is used in the retina}
\usage{
cortical_terms_context(term, retina_name = "en_associative",
  start_index = 0, max_results = 10, get_fingerprint = FALSE,
  api_key = NULL)
}
\arguments{
\item{term}{the search word that you would like to search for}

\item{retina_name}{one of en_associative or en_synomous see retina.}

\item{start_index}{start of the index (if multiple pages are returned)}

\item{max_results}{maximum number of result per page.}

\item{get_fingerprint}{true or false, return sparse matrix representation}

\item{api_key}{your key, if you saved the key in .renviron leave this NULL.}
}
\value{
a description of the context of this word in the retina.
}
\description{
Describes what contexts this word is used in.
}
\seealso{
Other terms: \code{\link{cortical_similar_terms}},
  \code{\link{cortical_terms}}
}

