% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{cortical_text_keywords}
\alias{cortical_text_keywords}
\title{Get a list of keywords that represent the text according to Cortical.io}
\usage{
cortical_text_keywords(text, retina_name = "en_associative", api_key = NULL)
}
\arguments{
\item{text}{input a character vector:  "a small lamb walked ..."}

\item{retina_name}{one of en_associative or en_synomous see retina.}

\item{api_key}{your key, if you saved the key in .renviron leave this NULL.}
}
\value{
a list of keywords
}
\description{
Get a list of keywords that represent the text according to Cortical.io
}
\seealso{
Other text: \code{\link{cortical_text_representation}},
  \code{\link{cortical_text_slices}},
  \code{\link{cortical_text_tokenize}}
}

