% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{cortical_text_language}
\alias{cortical_text_language}
\title{Detect language of a text}
\usage{
cortical_text_language(text, api_key = NULL)
}
\arguments{
\item{text}{input a character vector:  "a small lamb walked ..."}

\item{api}{your key, if you saved the key in .renviron leave this NULL.}
}
\value{
a list of language, iso tag and possible link to wikipedia
}
\description{
Post a piece of text to cortical.io to detect the language.
This endpoint will detect the language of an input text. It is capable
of identifying more than 50 languages. For best results it is recommended
that input texts consist of a minimum of 10 words (approximately 40 to 50
characters).
\url{http://api.cortical.io/Text.htm#!/text/getLanguage_post_5}
}
\seealso{
Other text: \code{\link{cortical_text_keywords}},
  \code{\link{cortical_text_representation}},
  \code{\link{cortical_text_slices}},
  \code{\link{cortical_text_tokenize}}
}
