% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{cortical_text_slices}
\alias{cortical_text_slices}
\title{Slice the text up}
\usage{
cortical_text_slices(text, retina_name = "en_associative", start_index = 0,
  max_results = 10, get_fingerprint = FALSE, api_key = NULL)
}
\arguments{
\item{text}{input a character vector:  "a small lamb walked ..."}

\item{retina_name}{one of en_associative or en_synomous see retina.}

\item{start_index}{start of the index (if multiple pages are returned)}

\item{max_results}{maximum number of result per page.}

\item{get_fingerprint}{true or false, return sparse matrix representation}

\item{api_key}{your key, if you saved the key in .renviron leave this NULL.}
}
\value{
results
}
\description{
Returns an ordered list of text objects (ordered according to where the text
slice appears in the input text). A Text object consists of a text slice
(defined as a slice by the Retina) and a Fingerprint object corresponding
to the text slice.
}
\seealso{
Other text: \code{\link{cortical_text_keywords}},
  \code{\link{cortical_text_language}},
  \code{\link{cortical_text_representation}},
  \code{\link{cortical_text_tokenize}}
}
