% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{cortical_text_tokenize}
\alias{cortical_text_tokenize}
\title{Tokenize the words or return only the part of speech (POS) parts}
\usage{
cortical_text_tokenize(text, pos_tags = NULL,
  retina_name = "en_associative", api_key = NULL)
}
\arguments{
\item{text}{input a character vector:  "a small lamb walked ..."}

\item{pos_tags}{a character representation such as "NNS" for nouns. see the link.}

\item{retina_name}{one of en_associative or en_synomous see retina.}

\item{api_key}{your key, if you saved the key in .renviron leave this NULL.}
}
\value{
seperate words or only the part of speech you asked for
}
\description{
Send a text and return seperate words or only nouns, verbs,adjectives, etc
or a combination.
\url{http://documentation.cortical.io/working_with_text.html#pos-tags}
}
\seealso{
Other text: \code{\link{cortical_text_keywords}},
  \code{\link{cortical_text_language}},
  \code{\link{cortical_text_representation}},
  \code{\link{cortical_text_slices}}
}
